requirements.txt
# Core RAG dependencies
openai>=1.0.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0

# Document parsing
pdfminer.six>=20221105
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Config + utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Evaluation
rouge-score>=0.1.2
